+++ src/main/kotlin/kr/io/snuhbmilab/carediaryserverv2/admin/facade/AdminDiaryFacade.kt
@@ -1,9 +1,7 @@
 package kr.io.snuhbmilab.carediaryserverv2.admin.facade
 
 import kr.io.snuhbmilab.carediaryserverv2.admin.dto.response.AdminDiaryFindAllResponse
-import kr.io.snuhbmilab.carediaryserverv2.admin.dto.response.AdminDiaryKeywordResponse
 import kr.io.snuhbmilab.carediaryserverv2.admin.dto.response.AdminDiarySdohResponse
-import kr.io.snuhbmilab.carediaryserverv2.admin.dto.response.AdminDiaryWelfareServiceResponse
 import kr.io.snuhbmilab.carediaryserverv2.domain.diary.service.DiaryAnalysisResultService
 import kr.io.snuhbmilab.carediaryserverv2.domain.diary.service.DiaryRecommendedQuestionService
 import kr.io.snuhbmilab.carediaryserverv2.domain.diary.service.DiaryService
@@ -28,27 +26,15 @@ class AdminDiaryFacade(
         return AdminDiarySdohResponse.from(pieList)
     }
 
-    fun findExtractedKeywords(diaryId: UUID): AdminDiaryKeywordResponse {
-        diaryService.validateExists(diaryId)
-        val keywords = diaryAnalysisResultService.findAllKeywordExtractions(diaryId)
-
-        return AdminDiaryKeywordResponse.of(diaryId, keywords)
-    }
-
-    fun findWelfareServices(diaryId: UUID): AdminDiaryWelfareServiceResponse {
-        diaryService.validateExists(diaryId)
-        val welfareServices = diaryAnalysisResultService.findAllWelfareServices(diaryId)
-
-        return AdminDiaryWelfareServiceResponse.of(diaryId, welfareServices)
-    }
-
     fun findAllByUserIdAndDate(userId: UUID, date: LocalDate): AdminDiaryFindAllResponse {
         userService.validateExists(userId)
         val diaries = diaryService.findAllByUserIdAndDate(userId, date)
         val diaryIds = diaries.mapNotNull { it.id }
         val questionScoresMap = diaryRecommendedQuestionService.findAllMapByDiaryIds(diaryIds)
+        val keywordExtractions = diaryAnalysisResultService.findKeywordExtractionsMapByDiaryIds(diaryIds)
+        val welfareServices = diaryAnalysisResultService.findWelfareServicesMapByDiaryIds(diaryIds)
 
-        return AdminDiaryFindAllResponse.of(diaries, questionScoresMap)
+        return AdminDiaryFindAllResponse.of(diaries, questionScoresMap, keywordExtractions, welfareServices)
     }
 
     @Transactional
