+++ src/main/kotlin/kr/io/snuhbmilab/carediaryserverv2/domain/diary/repository/DiaryKeywordExtractionRepository.kt
@@ -1,9 +1,11 @@
 package kr.io.snuhbmilab.carediaryserverv2.domain.diary.repository
 
+import kr.io.snuhbmilab.carediaryserverv2.domain.diary.entity.Diary
 import kr.io.snuhbmilab.carediaryserverv2.domain.diary.entity.DiaryKeywordExtraction
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.UUID
 
 interface DiaryKeywordExtractionRepository : JpaRepository<DiaryKeywordExtraction, Long> {
     fun findAllByDiaryId(diaryId: UUID): List<DiaryKeywordExtraction>
+    fun findAllByDiaryIdIn(diaryIds: List<UUID>): List<DiaryKeywordExtraction>
 }
